debug

+++ compass_pid.c
@@ -61,9 +61,9 @@ int main()
 		for(i=0; i<20; i++)
 		{
 			compass_angle = cal_compass_theta();
-			pid_data.input = compass_angle/180*100;
+			pid_data.input = (int)(compass_angle/180*100);
 			compute_output(&pid_data);
-			printf("pid_output = %f\n",pid_data.output);
+			printf("pid_output = %d\n",pid_data.output);
 			motor_slalom(-pid_data.output);
 			delay(50);
 		}
